requirements.txt
cbor2==5.2.0
Deprecated==1.2.10
luma.core==2.2.0
luma.led-matrix==1.5.0
paho-mqtt==1.5.1
Pillow==8.0.1
pyftdi==0.52.0
pyserial==3.5
pyusb==1.1.0
smbus2==0.4.0
wrapt==1.12.1
